+++ client/platform/web-girder/api/viameDetection.service.ts
