+++ client/platform/web-girder/main.ts
@@ -8,9 +8,9 @@ import { Vue as SentryVue } from '@sentry/integrations';
 import snackbarService from 'viame-web-common/vue-utilities/snackbar-service';
 import promptService from 'viame-web-common/vue-utilities/prompt-service';
 import vMousetrap from 'viame-web-common/vue-utilities/v-mousetrap';
-import vuetify from 'viame-web-common/plugins/vuetify';
-import girderRest from 'viame-web-common/plugins/girder';
 
+import vuetify from './plugins/vuetify';
+import girderRest from './plugins/girder';
 import App from './App.vue';
 import router from './router';
 import store from './store';
