+++ client/platform/web-girder/router.js
@@ -1,12 +1,13 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 
-import girderRest from 'viame-web-common/plugins/girder';
-import Viewer from 'viame-web-common/views/TrackViewer/Viewer.vue';
-import Home from 'viame-web-common/views/Home.vue';
-import Jobs from 'viame-web-common/views/Jobs.vue';
-import Login from 'viame-web-common/views/Login.vue';
-import Settings from 'viame-web-common/views/Settings.vue';
+import girderRest from './plugins/girder';
+
+import Home from './views/Home.vue';
+import Jobs from './views/Jobs.vue';
+import Login from './views/Login.vue';
+import Settings from './views/Settings.vue';
+import ViewerLoader from './views/ViewerLoader.vue';
 
 Vue.use(Router);
 
@@ -40,7 +41,7 @@ export default new Router({
     {
       path: '/viewer/:datasetId?',
       name: 'viewer',
-      component: Viewer,
+      component: ViewerLoader,
       props: true,
       beforeEnter,
     },
