+++ client/platform/web-girder/store/Pipelines.ts
