+++ client/platform/web-girder/views/Settings.vue
