+++ client/viame-web-common/components/Sidebar.vue
