+++ client/viame-web-common/use/index.ts
@@ -1,10 +1,8 @@
-import useGirderDataset from './useGirderDataset';
 import useModeManager from './useModeManager';
 import useSave from './useSave';
 import useSettings from './useSettings';
 
 export {
-  useGirderDataset,
   useModeManager,
   useSave,
   useSettings,
